first shot at splitting servant into servant, servant-client and servant-docs
