+++ servant-docs/servant-docs.cabal
@@ -1,0 +1,37 @@
+name:                servant-docs
+version:             0.2
+-- synopsis:
+-- description:
+license:             BSD3
+license-file:        LICENSE
+author:              Alp Mestanogullari, Sönke Hahn, Julian K. Arni
+maintainer:          alpmestan@gmail.com
+copyright:           2014 Zalora South East Asia Pte Ltd
+category:            Web
+build-type:          Simple
+cabal-version:       >=1.10
+tested-with:         GHC >= 7.8
+
+library
+  exposed-modules:
+    Servant.Docs
+  build-depends:
+      base >=4.7 && <5
+    , aeson
+    , bytestring
+    , hashable
+    , lens
+    , servant
+    , string-conversions
+    , system-filepath
+    , unordered-containers
+  hs-source-dirs: src
+  default-language: Haskell2010
+  ghc-options: -O0 -Wall
+
+executable greet-docs
+  main-is: greet.hs
+  hs-source-dirs: example
+  ghc-options: -O0 -Wall
+  build-depends: base, aeson, servant, servant-docs, text
+  default-language: Haskell2010
