+++ servant/example/greet.hs
@@ -3,13 +3,8 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE QuasiQuotes #-}
-{-# OPTIONS_GHC -fno-warn-orphans #-}
 
-import Control.Concurrent (forkIO, killThread)
-import Control.Monad.Trans.Either
 import Data.Aeson
 import Data.Monoid
 import Data.Proxy
@@ -18,10 +13,7 @@ import GHC.Generics
 import Network.Wai
 import Network.Wai.Handler.Warp
 
-import Servant.API
-import Servant.Client
-import Servant.Docs
-import Servant.Server
+import Servant
 
 -- * Example
 
@@ -31,25 +23,6 @@ newtype Greet = Greet { msg :: Text }
 
 instance FromJSON Greet
 instance ToJSON Greet
-
--- We add some useful annotations to our captures,
--- query parameters and request body to make the docs
--- really helpful.
-instance ToCapture (Capture "name" Text) where
-  toCapture _ = DocCapture "name" "name of the person to greet"
-
-instance ToCapture (Capture "greetid" Text) where
-  toCapture _ = DocCapture "greetid" "identifier of the greet msg to remove"
-
-instance ToParam (QueryParam "capital" Bool) where
-  toParam _ =
-    DocQueryParam "capital"
-                  ["true", "false"]
-                  "Get the greeting message in uppercase (true) or not (false). Default is false."
-                  Normal
-
-instance ToSample Greet where
-  toSample = Just $ Greet "Hello, haskeller!"
 
 -- API specification
 type TestApi =
@@ -83,29 +56,10 @@ server = helloH :<|> postGreetH :<|> deleteGreetH
 
         deleteGreetH _ = return ()
 
--- Client-side querying functions
---
--- They're all derived automatically from the type, and glued together
--- with :<|> just like in the type and for the server handlers, except
--- that we don't have to implement them!
-clientApi :: Client TestApi
-clientApi = client testApi
-
-getGreet :: Text -> Maybe Bool -> BaseUrl -> EitherT String IO Greet
-postGreet :: Greet -> BaseUrl -> EitherT String IO Greet
-deleteGreet :: Text -> BaseUrl -> EitherT String IO ()
-getGreet :<|> postGreet :<|> deleteGreet = clientApi
-
 -- Turn the server into a WAI app. 'serve' is provided by servant,
 -- more precisely by the Servant.Server module.
 test :: Application
 test = serve testApi server
-
--- Generate the data that lets us have API docs. This
--- is derived from the type as well as from
--- the 'ToCapture', 'ToParam' and 'ToSample' instances from above.
-docsGreet :: API
-docsGreet = docs testApi
 
 -- Run the server.
 --
@@ -115,21 +69,4 @@ runTestServer port = run port test
 
 -- Put this all to work!
 main :: IO ()
-main = do
-  -- we start the server, binding it to port 8001
-  tid <- forkIO $ runTestServer 8001
-
-  -- we tell the client where to find it
-  let uri = BaseUrl Http "localhost" 8001
-
-  -- we run a couple of requests against the server
-  print =<< runEitherT (getGreet "alp" (Just True) uri)
-  print =<< runEitherT (getGreet "alp" (Just False) uri)
-  let g = Greet "yo"
-  print =<< runEitherT (postGreet g uri)
-  print =<< runEitherT (deleteGreet "blah" uri)
-  killThread tid
-  putStrLn "\n---------\n"
-
-  -- we print the markdown docs
-  putStrLn $ markdown docsGreet
+main = runTestServer 8001
