+++ servant/src/Servant/API/Alternative.hs
@@ -3,9 +3,8 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 module Servant.API.Alternative where
 
+import Data.Monoid
 import Data.Proxy
-import Servant.Client
-import Servant.Docs
 import Servant.Server
 
 -- | Union of two APIs, first takes precedence in case of overlap.
@@ -38,35 +37,3 @@ instance (HasServer a, HasServer b) => HasServer (a :<|> b) where
 
     where pa = Proxy :: Proxy a
           pb = Proxy :: Proxy b
-
--- | A client querying function for @a ':<|>' b@ will actually hand you
---   one function for querying @a@ and another one for querying @b@,
---   stitching them together with ':<|>', which really is just like a pair.
---
--- > type MyApi = "books" :> Get [Book] -- GET /books
--- >         :<|> "books" :> ReqBody Book :> Post Book -- POST /books
--- >
--- > myApi :: Proxy MyApi
--- > myApi = Proxy
--- >
--- > getAllBooks :: BaseUrl -> EitherT String IO [Book]
--- > postNewBook :: Book -> BaseUrl -> EitherT String IO Book
--- > (getAllBooks :<|> postNewBook) = client myApi
-instance (HasClient a, HasClient b) => HasClient (a :<|> b) where
-  type Client (a :<|> b) = Client a :<|> Client b
-  clientWithRoute Proxy req =
-    clientWithRoute (Proxy :: Proxy a) req :<|>
-    clientWithRoute (Proxy :: Proxy b) req
-
--- | The generated docs for @a ':<|>' b@ just appends the docs
---   for @a@ with the docs for @b@.
-instance (HasDocs layout1, HasDocs layout2)
-      => HasDocs (layout1 :<|> layout2) where
-
-  docsFor Proxy (ep, action) = docsFor p1 (ep, action) <> docsFor p2 (ep, action)
-
-    where p1 :: Proxy layout1
-          p1 = Proxy
-
-          p2 :: Proxy layout2
-          p2 = Proxy
