+++ servant/src/Servant/API/Capture.hs
@@ -11,10 +11,7 @@ import Data.Text
 import GHC.TypeLits
 import Network.Wai
 import Servant.API.Sub
-import Servant.Client
-import Servant.Common.Req
 import Servant.Common.Text
-import Servant.Docs
 import Servant.Server
 
 -- | Capture a value from the request path under a certain type @a@.
@@ -61,49 +58,3 @@ instance (KnownSymbol capture, FromText a, HasServer sublayout)
     _ -> respond $ failWith NotFound
 
     where captureProxy = Proxy :: Proxy (Capture capture a)
-
--- | If you use a 'Capture' in one of your endpoints in your API,
--- the corresponding querying function will automatically take
--- an additional argument of the type specified by your 'Capture'.
--- That function will take care of inserting a textual representation
--- of this value at the right place in the request path.
---
--- You can control how values for this type are turned into
--- text by specifying a 'ToText' instance for your type.
---
--- Example:
---
--- > type MyApi = "books" :> Capture "isbn" Text :> Get Book
--- >
--- > myApi :: Proxy MyApi
--- > myApi = Proxy
--- >
--- > getBook :: Text -> BaseUrl -> EitherT String IO Book
--- > getBook = client myApi
--- > -- then you can just use "getBook" to query that endpoint
-instance (KnownSymbol capture, ToText a, HasClient sublayout)
-      => HasClient (Capture capture a :> sublayout) where
-
-  type Client (Capture capture a :> sublayout) =
-    a -> Client sublayout
-
-  clientWithRoute Proxy req val =
-    clientWithRoute (Proxy :: Proxy sublayout) $
-      appendToPath p req
-
-    where p = unpack (toText val)
-
--- | @"books" :> 'Capture' "isbn" Text@ will appear as
--- @/books/:isbn@ in the docs.
-instance (KnownSymbol sym, ToCapture (Capture sym a), HasDocs sublayout)
-      => HasDocs (Capture sym a :> sublayout) where
-
-  docsFor Proxy (endpoint, action) =
-    docsFor sublayoutP (endpoint', action')
-
-    where sublayoutP = Proxy :: Proxy sublayout
-          captureP = Proxy :: Proxy (Capture sym a)
-
-          action' = over captures (|> toCapture captureP) action
-          endpoint' = over path (\p -> p++"/:"++symbolVal symP) endpoint
-          symP = Proxy :: Proxy sym
