+++ servant/src/Servant/API/Put.hs
@@ -11,9 +11,6 @@ import Data.String.Conversions
 import Data.Typeable
 import Network.HTTP.Types
 import Network.Wai
-import Servant.Client
-import Servant.Common.Req
-import Servant.Docs
 import Servant.Server
 
 -- | Endpoint for PUT requests, usually used to update a ressource.
@@ -53,24 +50,3 @@ instance ToJSON a => HasServer (Put a) where
         respond $ failWith WrongMethod
 
     | otherwise = respond $ failWith NotFound
-
--- | If you have a 'Put' endpoint in your API, the client
--- side querying function that is created when calling 'client'
--- will just require an argument that specifies the scheme, host
--- and port to send the request to.
-instance FromJSON a => HasClient (Put a) where
-  type Client (Put a) = BaseUrl -> EitherT String IO a
-
-  clientWithRoute Proxy req host =
-    performRequestJSON methodPut req 200 host
-
-instance ToSample a => HasDocs (Put a) where
-  docsFor Proxy (endpoint, action) =
-    single endpoint' action'
-
-    where endpoint' = endpoint & method .~ DocPUT
-
-          action' = action & response.respBody .~ sampleByteString p
-                           & response.respStatus .~ 200
-
-          p = Proxy :: Proxy a
